Output to UART Success
Program successfully performs arithmetic operations, & outputs to the
UART with the output file given on Blackboard.

Program reads input, determines if number, otherwise checks if
operation, when it gets to check operation it determines if there was a
operation already beforehand (if there is it will perform the operation
then return to determine the new operation), else determine new
operation and get next input.

If we read a equal sign, we perform operation if there is one,
otherwise we output to UART and reset all values.